Add feature where COVID period can be specified in the data
To specify a COVID period different from 1/1/2020 to 5/5/2023, add the word "START" to the column which is the first period subject of COVID in the row above the period row. Also add the word "END" to the column which is the last period subject of COVID in the row above the period row.

The program will use this as the COVID period in its computations.